requirements.txt
.
tensorflow_addons
pylint
sphinx
furo
